+++ Tuist/ProjectDescriptionHelpers/Module.swift
@@ -413,6 +413,7 @@ public enum Module: String, CaseIterable {
                     .target(name: Module.testCommand.targetName),
                     .external(name: "FileSystem"),
                     .external(name: "FileSystemTesting"),
+                    .external(name: "XcodeProj"),
                 ]
             case .dependencies:
                 [
@@ -735,6 +736,7 @@ public enum Module: String, CaseIterable {
                     .external(name: "XCLogParser"),
                     .external(name: "OrderedSet"),
                     .external(name: "Crypto"),
+                    .external(name: "SwifterPMCore"),
                 ]
             case .kit:
                 [
@@ -1648,6 +1650,7 @@ public enum Module: String, CaseIterable {
                     .external(name: "FileSystem"),
                     .external(name: "FileSystemTesting"),
                     .external(name: "Command"),
+                    .external(name: "SwifterPMCore"),
                 ]
             case .projectDescription:
                 [